+++ com.unity.ml-agents/Runtime/Agent.cs
@@ -23,7 +23,7 @@ namespace MLAgents
         /// For discrete control, specifies the actions that the agent cannot take. Is true if
         /// the action is masked.
         /// </summary>
-        public bool[] actionMasks;
+        public bool[] discreteActionMasks;
 
         /// <summary>
         /// Current agent reward.
@@ -117,6 +117,19 @@ namespace MLAgents
         internal struct AgentParameters
         {
             public int maxStep;
+        }
+
+        public int TeamId {
+            get {
+                LazyInitialize();
+                return m_PolicyFactory.TeamId;
+                }
+        }
+        public string BehaviorName  {
+            get {
+                LazyInitialize();
+                return m_PolicyFactory.behaviorName;
+                }
         }
 
         [SerializeField][HideInInspector]
@@ -351,6 +364,22 @@ namespace MLAgents
         }
 
         /// <summary>
+        /// Updates the type of behavior for the agent.
+        /// </summary>
+        /// <param name="behaviorType"> The new behaviorType for the Agent
+        /// </param>
+        public void SetBehaviorType(BehaviorType behaviorType)
+        {
+            if (m_PolicyFactory.m_BehaviorType == behaviorType)
+            {
+                return;
+            }
+            m_PolicyFactory.m_BehaviorType = behaviorType;
+            m_Brain?.Dispose();
+            m_Brain = m_PolicyFactory.GeneratePolicy(Heuristic);
+        }
+
+        /// <summary>
         /// Returns the current step counter (within the current episode).
         /// </summary>
         /// <returns>
@@ -442,16 +471,8 @@ namespace MLAgents
             // should stay the previous action before the Done(), so that it is properly recorded.
             if (m_Action.vectorActions == null)
             {
-                if (param.vectorActionSpaceType == SpaceType.Continuous)
-                {
-                    m_Action.vectorActions = new float[param.vectorActionSize[0]];
-                    m_Info.storedVectorActions = new float[param.vectorActionSize[0]];
-                }
-                else
-                {
-                    m_Action.vectorActions = new float[param.vectorActionSize.Length];
-                    m_Info.storedVectorActions = new float[param.vectorActionSize.Length];
-                }
+                m_Action.vectorActions = new float[param.numActions];
+                m_Info.storedVectorActions = new float[param.numActions];
             }
         }
 
@@ -477,13 +498,10 @@ namespace MLAgents
         /// </returns>
         public virtual float[] Heuristic()
         {
-            Debug.LogWarning("Heuristic method called but not implemented. Return placeholder actions.");
+            Debug.LogWarning("Heuristic method called but not implemented. Returning placeholder actions.");
             var param = m_PolicyFactory.brainParameters;
-            var actionSize = param.vectorActionSpaceType == SpaceType.Continuous ?
-                param.vectorActionSize[0] :
-                param.vectorActionSize.Length;
 
-            return new float[actionSize];
+            return new float[param.numActions];
         }
 
         /// <summary>
@@ -564,7 +582,7 @@ namespace MLAgents
                     CollectDiscreteActionMasks(m_ActionMasker);
                 }
             }
-            m_Info.actionMasks = m_ActionMasker.GetMask();
+            m_Info.discreteActionMasks = m_ActionMasker.GetMask();
 
             m_Info.reward = m_Reward;
             m_Info.done = false;